requirements.txt
fastapi
uvicorn[standard]
fpdf
python-dotenv
pydantic


